// __generated__/api-client/model/inlineResponse2002.ts
/**
 * CryptoAPIs
 * Crypto APIs 2.0 is a complex and innovative infrastructure layer that radically simplifies the development of any Blockchain and Crypto related applications. Organized around REST, Crypto APIs 2.0 can assist both novice Bitcoin/Ethereum enthusiasts and crypto experts with the development of their blockchain applications. Crypto APIs 2.0 provides unified endpoints and data, raw data, automatic tokens and coins forwardings, callback functionalities, and much more.
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: developers@cryptoapis.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InlineResponse2002Data } from "./inlineResponse2002Data"

export interface InlineResponse2002 {
  /**
   * Specifies the version of the API that incorporates this endpoint.
   */
  apiVersion: string
  /**
   * Context in every response will be \'mobile\'
   */
  context: string
  data: InlineResponse2002Data
}